+++ CHANGELOG.md
@@ -8,7 +8,8 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.0.7] - TBD
 ### Changed
 - Deprecate Uuid.parse() in favor of Uuid.fromString(), which returns a non-null Uuid or throws an error for an invalid string, in line with Java's UUID.fromString().
-- `uuidOf(msb: Long, lsb: Long)` function (#33)
+- `uuidOf(msb: Long, lsb: Long)` function (#52)
+- Deprecated `Uuid(msb: Long, lsb: Long)` in favor of `uuidOf` (#64)
 
 ## [0.0.6] - 2019-11-21
 ### Changed